+++ src/LinkChildTagHelper.cs
@@ -1,5 +1,7 @@
 namespace Tailwind.Css.TagHelpers;
 
+using System.Text.Encodings.Web;
+
 using Microsoft.AspNetCore.Razor.TagHelpers;
 using Microsoft.Extensions.Options;
 
@@ -7,8 +9,10 @@ using Microsoft.Extensions.Options;
 [HtmlTargetElement(ParentTag = "a", Attributes = DefaultClassAttributeName)]
 public class LinkChildTagHelper : LinkTagHelperBase
 {
-    public LinkChildTagHelper(IOptions<TagOptions> settings)
-        : base(settings)
+    public LinkChildTagHelper(
+        IOptions<TagOptions> settings,
+        HtmlEncoder htmlEncoder)
+        : base(settings, htmlEncoder)
     {
     }