+++ src/Utilities.cs
@@ -1,10 +1,35 @@
 namespace Tailwind.Css.TagHelpers;
 
 using System;
+using System.Text.Encodings.Web;
+
+using Microsoft.AspNetCore.Html;
+using Microsoft.AspNetCore.Razor.TagHelpers;
 
 internal static class Utilities
 {
     private static readonly char[] SpaceChars = { '\u0020', '\u0009', '\u000A', '\u000C', '\u000D' };
+
+    public static string ExtractClassValue(TagHelperOutput output, HtmlEncoder encoder)
+    {
+        if (!output.Attributes.TryGetAttribute("class", out var classAttribute))
+        {
+            return "";
+        }
+
+        if (classAttribute?.Value is null)
+        {
+            return "";
+        }
+
+        return classAttribute.Value switch
+        {
+            string stringValue => stringValue,
+            HtmlString htmlString => htmlString.Value,
+            IHtmlContent htmlContent => ExtractHtmlContent(encoder, htmlContent),
+            _ => classAttribute.Value.ToString() ?? "",
+        };
+    }
 
     public static string[]? SplitClassList(string? classes)
     {
@@ -16,5 +41,14 @@ internal static class Utilities
         return classes.Split(
             SpaceChars,
             StringSplitOptions.RemoveEmptyEntries);
+    }
+
+    private static string ExtractHtmlContent(HtmlEncoder encoder, IHtmlContent htmlContent)
+    {
+        using var stringWriter = new StringWriter();
+
+        htmlContent.WriteTo(stringWriter, encoder);
+
+        return stringWriter.ToString();
     }
 }